Version 0.2.2 — playlists intelligentes, colonnes au choix, export
Playlists intelligentes : des règles sur 24 champs de la bibliothèque
et sur n'importe quel tag des fichiers, groupes imbriqués, tri et coupe.
Le nombre de morceaux retenus se recalcule pendant qu'on écrit les règles.
Huit recettes toutes faites, dont « les styles les plus écoutés », qui
somme les écoutes par genre au lieu de figer une liste.

Compteur d'écoutes : il n'existait pas. La fonction était écrite mais
jamais appelée, et play_count valait zéro sur toute la bibliothèque. Une
écoute compte au-delà de la moitié du morceau, plafonnée à une minute.

Colonnes au choix dans les listes de pistes : 21 champs de la
bibliothèque et tous les tags réellement présents dans les fichiers,
recensés avec leur effectif. Tri par en-tête sur n'importe lequel, dans
les deux sens — en base pour l'onglet Morceaux, en mémoire ailleurs.
La vue tableau s'étend aux albums, artistes, genres et playlists.

Export et import des réglages, profils, playlists et titres aimés. Les
playlists n'emportent pas d'identifiants, qui n'auraient aucun sens
ailleurs : chaque piste part avec son chemin et de quoi la reconnaître
si ce chemin a changé. L'import annonce ce qu'il fera avant de l'écrire.

Notation en demi-étoiles, en décimal — les notes existantes valent déjà
1,0 à 5,0, aucune donnée n'est transformée.

Corrections
- La file ne s'enchaînait pas en bit-perfect : le Drop des sorties
  exclusives levait le drapeau d'arrêt, et playback-ended n'était jamais
  émis. WASAPI et ALSA étaient touchés.
- Les compteurs de playlist mentaient. Retirer un dossier efface en
  cascade fichiers, pistes et entrées de playlist ; le compteur restait
  sur son ancienne valeur. Il se lit désormais au lieu d'être cru.
- Défilement absent sur les trois pages de playlist.
- Menus déroulants blancs en thème sombre : color-scheme n'était pas
  déclaré, et un fond translucide ne peut pas servir de surface à une
  liste déroulante.
- Listes plus fluides : content-visibility sur les lignes, et un flou
  d'arrière-plan par piste supprimé — son fond était uni.
- Le téléchargement des portraits d'artistes est débrayable.
- La fenêtre retient sa taille et sa position.

Notes techniques
- Les tags des fichiers remontent jusqu'à l'interface, qui y puise les
  colonnes choisies.
- Aucun nom de champ ni valeur de règle n'entre dans le texte d'une
  requête : tout part en valeur liée, y compris les noms de tags, qui
  viennent des fichiers.
- 285 tests, dont une trentaine qui exécutent le SQL contre une base
  migrée — sqlx apparie colonnes et champs à l'exécution, une requête
  incomplète compile sans broncher.

+++ src-tauri/src/entity/audio/audio_tags.rs
@@ -12,7 +12,7 @@ pub struct AudioTags {
     pub comment: Option<String>,
     pub track_number: Option<u16>,
     pub genre: Option<String>,
-    pub rating: Option<i32>, // 0-5 stars (POPM → mappé)
+    pub rating: Option<f64>, // 0,5 à 5,0 par demi (POPM → mappé)
 
     // Extensions for ID3v2 and other formats
     pub album_artist: Option<String>,
