+++ src-tauri/src/entity/library/library_track.rs
@@ -25,7 +25,7 @@ pub struct LibraryTrack {
     
     pub play_count: i64,
     pub last_played_at: Option<DateTime<Utc>>,
-    pub rating: Option<i32>,
+    pub rating: Option<f64>,
     pub favorite: bool,
     
     pub created_at: DateTime<Utc>,
@@ -51,5 +51,5 @@ pub struct LibraryTrackCreate {
     pub duration: Option<f64>,
     pub bitrate: Option<i32>,
     pub sample_rate: Option<i32>,
-    pub rating: Option<i32>,
+    pub rating: Option<f64>,
 }
